1.修改删除用户只有订阅状态为false时可删 2.将去除词的TXT放在根目录下 3.修改删除管理员时跳转不正确 4.删除操作流水时将user表的filename字段置为空

+++ src/main/java/com/book/web/UserController.java
@@ -41,17 +41,23 @@ public class UserController {
     @RequestMapping("/deleteUser.html")
     public String deleteUser(HttpServletRequest request,RedirectAttributes redirectAttributes){
         String userId = request.getParameter("userId");
+        String isSubscribe = request.getParameter("isSubscribe");
         String filePath = "D:\\apache-tomcat-7.0.84\\bin\\chatData\\" + userId +".txt";
         File file = new File(filePath);
         file.delete();
-        int success=userService.deleteUserList(userId);
-        if(success==1){
-            redirectAttributes.addFlashAttribute("succ", "删除成功！");
-            return "redirect:/userlist.html";
-        }else {
-            redirectAttributes.addFlashAttribute("error", "删除失败！");
-            return "redirect:/userlist.html";
+
+        if(isSubscribe.equals("false")){
+            int success=userService.deleteUserList(userId);
+            if(success==1){
+                redirectAttributes.addFlashAttribute("succ", "删除成功！");
+                return "redirect:/userlist.html";
+            }else {
+                redirectAttributes.addFlashAttribute("error", "删除失败！");
+                return "redirect:/userlist.html";
+            }
         }
+        redirectAttributes.addFlashAttribute("error", "不能删除订阅用户！");
+        return "redirect:/userlist.html";
     }
     @RequestMapping("/queryUser.html")
     public ModelAndView queryUser(HttpServletRequest request,RedirectAttributes redirectAttributes) throws UnsupportedEncodingException {
